Add custom hook for appending to typewrite input and reactivate insertion of assumptions into tactics via click.
