+++ client/src/components/editor/typewriter/proof_step/Hyp.tsx
@@ -1,5 +1,6 @@
 import { InteractiveHypothesisBundle, InteractiveHypothesisBundle_nonAnonymousNames, MVarId, TaggedText_stripTags } from "@leanprover/infoview-api"
 import React from "react"
+import {useAppendTypewriterInput} from "../useAppendTypewriterInput"
 
 interface HypProps {
     hyp: InteractiveHypothesisBundle
@@ -11,19 +12,19 @@ function isInaccessibleName(h: string): boolean {
 }
 
 export function Hyp({ hyp: h, mvarId }: HypProps) {
-
     const namecls: string = 'mr1 hyp-name ' +
         (h.isInserted ? 'inserted-text ' : '') +
         (h.isRemoved ? 'removed-text ' : '')
 
-    // const appendTypewriterInput = useAppendTypewriterInput()
+    const appendTypewriterInput = useAppendTypewriterInput()
+
     const names = InteractiveHypothesisBundle_nonAnonymousNames(h).map((n, i) =>
         <span
             className={namecls + (isInaccessibleName(n) ? 'goal-inaccessible ' : '')}
             key={i}
             // FIXME: 3rd argument looks wrong
             onClick={ev => {
-                // appendTypewriterInput(ev.shiftKey, n, (h as any).isAssumption ?? false, (h as any).isAssumption ?? false)
+                appendTypewriterInput(ev.shiftKey, n, (h as any).isAssumption ?? false, (h as any).isAssumption ?? false)
                 ev.stopPropagation()
             } }
         >{n}
